dahdi install: if dahdi is not used do nothing

+++ pre-start.d/10-install-dahdi.sh
@@ -9,13 +9,18 @@ SENTINEL="/var/lib/wazo-upgrade/install-dahdi"
 
 [ -e "${SENTINEL}" ] && exit 0
 
-echo "Installing Dahdi dependencies"
-apt-get install wazo-asterisk-extra-modules
+# If DAHDI was already configured, reinstall it
+if [ -e "/etc/asterisk/dahdi-channels.conf" ]; then
 
-echo "Enabling Dahdi module"
-cat<<EOF > /etc/wazo-confgend/conf.d/50_enable_chan_dahdi.yml
+    echo "Installing DAHDI dependencies"
+    apt-get install wazo-asterisk-extra-modules
+
+    echo "Enabling DAHDI module"
+    cat << EOF > /etc/wazo-confgend/conf.d/50_enable_chan_dahdi.yml
 enabled_asterisk_modules:
   chan_dahdi.so: true
 EOF
+
+fi
 
 touch "${SENTINEL}"
